+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/psi/reference/PolyglotQueryDefinedTableReference.kt
@@ -23,6 +23,8 @@ import com.intellij.idea.plugin.hybris.psi.util.PsiUtils
 import com.intellij.idea.plugin.hybris.system.type.codeInsight.completion.TSCompletionService
 import com.intellij.idea.plugin.hybris.system.type.meta.TSMetaModelAccess
 import com.intellij.idea.plugin.hybris.system.type.meta.TSModificationTracker
+import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaEnum
+import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaItem
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaType
 import com.intellij.idea.plugin.hybris.system.type.psi.reference.result.EnumResolveResult
 import com.intellij.idea.plugin.hybris.system.type.psi.reference.result.ItemResolveResult
@@ -56,18 +58,20 @@ class PolyglotQueryDefinedTableReference(owner: PolyglotQueryTypeKeyName) : PsiR
         private val provider = ParameterizedCachedValueProvider<Array<ResolveResult>, PolyglotQueryDefinedTableReference> { ref ->
             val lookingForName = ref.element.typeName
             val project = ref.element.project
-            val modelAccess = TSMetaModelAccess.getInstance(project)
 
-            val result: Array<ResolveResult> = modelAccess.findMetaItemByName(lookingForName)
-                ?.declarations
-                ?.map { ItemResolveResult(it) }
+            val results: Array<ResolveResult> = TSMetaModelAccess.getInstance(project).findMetaClassifierByName(lookingForName)
+                ?.let {
+                    when (it) {
+                        is TSGlobalMetaItem -> it.declarations.map { meta -> ItemResolveResult(meta) }
+                        is TSGlobalMetaEnum -> it.declarations.map { meta -> EnumResolveResult(meta) }
+                        else -> null
+                    }
+                }
                 ?.toTypedArray()
-                ?: modelAccess.findMetaEnumByName(lookingForName)
-                    ?.let { arrayOf(EnumResolveResult(it)) }
                 ?: ResolveResult.EMPTY_ARRAY
 
             CachedValueProvider.Result.create(
-                result,
+                results,
                 project.service<TSModificationTracker>(), PsiModificationTracker.MODIFICATION_COUNT
             )
         }